Day 5 pt 2
This was a fun one! My first part felt brute force, so it was nice to
come up with a neater way of doing it by merging the ranges for part 2.
One thing that I miss with zig compared to Rust is more complex enum
definitions. I love that I can inline complex layered structs in Rust. I
also don't like having to do the `@This()` syntax with structs, or
having to deal with defining structs in the right order.